Add files via upload
Synchronized the xlsx file so that when you add, remove or edit courses, it does it in the xlsx file as well. Currently, only works in Courses and Subjects as they were the only ones who had the GUI set up to do so, but I will extend to other parts.